use transaction in handle_new_stream
Sometimes this function fails when a new stream is added which uses
games in the datapoints which are found in the database with the field
name set to the name in the datapoints, but where the twitch_name is
unset.
We try adding a new game with name=twitch_name=dp.game_name, which fails
the UNIQUEness check for name. Totally expected.

However, the problem now comes that since we weren't executing the
database operations in handle_new_stream in a transaction, we leave the
database in a half broken state.

This commit makes it such that all operations done in handle_new_stream
are using the same transaction, makeing it able to rollback everything
on any error.
This commit also requires some changes in db.rs, and files that call
functions in db.rs.