StreamDatapoint: serialize `game` field if not empty
This is required for streamwatch-tool to retain the `game` fields, and
in general it is more correct.

The exception is that we don't want to include the `game` field in the
serialized datapoints array in the database, since it is superfluous (we
already have the `game_features` table).
Before the only code that serialized the `StreamDatapoint` struct was
this database insertion code, so `#[serde(skip_serializing)]` on the
field was fine.

But now with streamwatch-tool this doesn't work anymore, and this commit
seems like the better solution:
We skip serializing the `game` field when it contains an empty string.
We serialize a version of the datapoints with every `game` field set to
an empty string when inserting a new stream in the database.