don't reuse possible_games over handle_new_stream invocations
The problem is that a clean transaction is used in every
handle_new_stream invocation, if this transaction is rolled back, there
is a possible mismatch between the database and possible_games.

This is currently not a problem, since when an error occurs we propagate
this error over scan_streams, thus stopping the new streams iteration,
but if this is ever changed this bug is uncovered.

There are ways to make this work somewhat efficiently, but
Database::get_possible_games is so cheap that I don't think it is
something to worry about.